backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
alembic>=1.13,<2
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
cryptography>=42,<44
pydantic-settings>=2.2,<3
pydantic[email]>=2.0,<3
rcon>=2.0,<3
slowapi>=0.1,<1
apscheduler>=3.10,<4
